+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# [v1.1.0](https://github.com/sitecite/extension/pull/2) [2026-06-28]
+
+* Optimised a lot of code
+    - Made it easier to convert browser formats
+    - Split off the `background.js` file into multiple different files
+
+* Updated installation instructions
+* Made the "Open sitecite" button clearer
+* Added symbols for the keyboard shortcut when using macOS
+
+
+## Bugs fixed
+
+* Fixed an issue where a success message did not appear when using a self-hosted instance
+* Fixed an issue with minumum font size on Chrome
+* Fixed an issue with the window size on Chrome
+
 # [v1.0.0](https://github.com/sitecite/extension/pull/1) [2026-6-15]
 
 * A link to the sitecite website is displayed more prominently